requirements.txt
# # # Core dependencies
# # fastapi==0.104.1
# # uvicorn==0.24.0
# # pydantic==2.5.1
# # python-dotenv==1.0.0

# # # Web scraping
# # beautifulsoup4==4.12.2
# # requests==2.31.0
# # selenium==4.15.2

# # # Google AI and embeddings
# # google-generativeai==0.3.0
# # google-cloud-aiplatform==1.38.1

# # # Vector database
# # pinecone-client==2.2.4

# # # Data processing
# # pandas==2.1.3
# # numpy==1.26.2

# # # Evaluation metrics
# # scikit-learn==1.3.2

# # # API client
# # httpx==0.25.1

# # # Web interface
# # streamlit==1.28.2

# # # Testing
# # pytest==7.4.3
# # pytest-asyncio==0.21.1

# # # Development tools
# # black==23.11.0
# # isort==5.12.0
# fastapi==0.104.1
# uvicorn==0.24.0
# pandas==2.1.3
# numpy==1.26.2
# python-dotenv==1.0.0
# requests==2.31.0
# beautifulsoup4==4.12.2
# streamlit==1.28.2
# plotly==5.18.0
# pydantic==2.5.2
# google-generativeai==0.3.2
# scikit-learn==1.3.2
# sentence-transformers==2.2.2
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
streamlit==1.28.2
plotly==5.18.0
pydantic==2.5.2
google-generativeai==0.3.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
transformers==4.36.0
huggingface-hub==0.19.4
torch==2.1.0